Add new option '--sccs-timezone'
Add new option '--sccs-timezone' for setting the timezone each SCCS
commit was made in.  Defaults to UTC.

Note that this assumes that all SCCS deltas were committed in the same
timezone.  Since older SCCS versions doesn't store timezone
information, it is _impossible_ to determine when each delta was
committed to any degree of accuracy[1].

Footnotes:
[1]  http://sccs.berlios.de/man/sccscvt.1.html